# e-com-nextjs/ferio-nest-prisma/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/admin/storage/presign-put": {
      "post": {
        "operationId": "StorageController_presignPut",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignPutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Storage"
        ]
      }
    },
...
    "/api/v1/admin/transactional-messages": {
      "get": {
        "operationId": "TransactionalMessagingController_getMessages",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
    "/api/v1/admin/transactional-messages/policy": {
      "get": {
        "operationId": "TransactionalMessagingController_getPolicy",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      },
      "patch": {
        "operationId": "TransactionalMessagingController_updatePolicy",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessagingPolicyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
    "/api/v1/admin/transactional-messages/templates": {
      "get": {
        "operationId": "TransactionalMessagingController_getTemplates",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
    "/api/v1/admin/transactional-messages/providers": {
      "get": {
        "operationId": "TransactionalMessagingController_getProviders",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
    "/api/v1/admin/transactional-messages/providers/{channel}": {
      "patch": {
        "operationId": "TransactionalMessagingController_updateProvider",
        "parameters": [
          {
            "name": "channel",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessagingProviderConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
    "/api/v1/admin/transactional-messages/templates/{key}": {
      "patch": {
        "operationId": "TransactionalMessagingController_updateTemplate",
        "parameters": [
          {
            "name": "key",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMessageTemplateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
...
    "/api/v1/admin/transactional-messages/{id}/retry": {
      "post": {
        "operationId": "TransactionalMessagingController_retry",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Admin Transactional Messages"
        ]
      }
    },
...
    "/api/v1/warranty/order-items": {
      "post": {
        "operationId": "WarrantyController_items",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyWarrantyOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Warranty"
        ]
      }
    },
...
    "/api/v1/warranty/claims": {
      "post": {
        "operationId": "WarrantyController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWarrantyClaimDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Warranty"
        ]
      }
    },
...
    "/api/v1/admin/warranty/{id}/status": {
      "patch": {
        "operationId": "AdminWarrantyController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarrantyClaimDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AdminWarranty"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ferio Commerce API",
    "description": "Ferio modular NestJS backend with PostgreSQL and Prisma",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "PresignPutDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateMessagingProviderConfigDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMessageTemplateDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateMessagingPolicyDto": {
        "type": "object",
        "properties": {}
      },
...
      "VerifyWarrantyOrderDto": {
        "type": "object",
        "properties": {}
      },
      "CreateWarrantyClaimDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateWarrantyClaimDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}